# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tags:
    get:
      operationId: listTags
      tags: [organize]
      summary: List tags
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTag
      tags: [organize]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tags/{tagId}:
    parameters:
      - name: tagId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateTag
      tags: [organize]
      summary: Update a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTag
      tags: [organize]
      summary: Delete a tag
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /custom-fields:
    get:
      operationId: listCustomFields
      tags: [organize]
      summary: List custom fields
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFieldList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCustomField
      tags: [organize]
      summary: Create a custom field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFieldInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /custom-fields/{fieldId}:
    parameters:
      - name: fieldId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateCustomField
      tags: [organize]
      summary: Update a custom field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFieldInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomField
      tags: [organize]
      summary: Delete a custom field
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mention-groups:
    get:
      operationId: listMentionGroups
      tags: [organize]
      summary: List mention groups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentionGroupList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMentionGroup
      tags: [organize]
      summary: Create a mention group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentionGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentionGroup"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mention-groups/{groupId}:
    parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateMentionGroup
      tags: [organize]
      summary: Update a mention group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentionGroupInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentionGroup"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMentionGroup
      tags: [organize]
      summary: Delete a mention group
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        postCount: { type: integer }
      required: [id, name, color]
    TagList:
      type: object
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
      required: [tags]
    TagInput:
      type: object
      properties:
        name: { type: string, maxLength: 60 }
        color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
    CustomField:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [text, number, date, select, url]
        options:
          type: array
          items: { type: string }
        required: { type: boolean }
        position: { type: integer }
      required: [id, key, label, type, options, required, position]
    CustomFieldList:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/CustomField"
      required: [fields]
    CustomFieldInput:
      type: object
      properties:
        label: { type: string, maxLength: 80 }
        type:
          type: string
          enum: [text, number, date, select, url]
        options:
          type: array
          items: { type: string }
        required: { type: boolean }
        position: { type: integer }
    MentionGroup:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        handles:
          type: array
          items: { type: string }
      required: [id, name, handles]
    MentionGroupList:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/MentionGroup"
      required: [groups]
    MentionGroupInput:
      type: object
      properties:
        name: { type: string, maxLength: 60 }
        handles:
          type: array
          items: { type: string }
...